Soucre Code/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
pandas==2.2.2
plotly==5.22.0
azure-storage-blob==12.24.0

